requirements.txt
streamlit==1.35.0
openai==1.30.1
pandas
python-dotenv
pycaret
langfuse==2.60.7